added physician rest controller, not found exception handling

+++ src/main/java/com/ntm/hospital/repository/PatientRepository.java
@@ -3,5 +3,5 @@ package com.ntm.hospital.repository;
 import com.ntm.hospital.data.Patient;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PatientRepository extends CrudRepository<Patient, Long> {
+public interface PatientRepository extends CrudRepository<Patient, Integer> {
 }
