+++ src/main/java/com/ntm/hospital/repository/PhysicianRepository.java
@@ -3,5 +3,5 @@ package com.ntm.hospital.repository;
 import com.ntm.hospital.data.Physician;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PhysicianRepository extends CrudRepository<Physician, Long> {
+public interface PhysicianRepository extends CrudRepository<Physician, Integer> {
 }
